tweak the guys bot to also handle other uninclusive language
