+++ package.json
@@ -40,10 +40,11 @@
   "devDependencies": {
     "@18f/18f-eslint": "^1.2.5",
     "chai": "^4.2.0",
+    "chai-subset": "^1.6.0",
     "co": "^4.6.0",
     "eslint-plugin-coffeescript": "^1.0.0",
-    "mocha": "^7.1.2",
     "hubot-test-helper": "^1.9.0",
+    "mocha": "^7.1.2",
     "nyc": "^15.0.1",
     "sinon": "^9.0.2"
   },
